refactor : add useHanldeImage
